+++ apache/tests/conftest.py
@@ -4,7 +4,6 @@
 
 import os
 
-import mock
 import pytest
 import requests
 
@@ -12,7 +11,7 @@ from datadog_checks.apache import Apache
 from datadog_checks.dev import docker_run
 from datadog_checks.dev.conditions import CheckEndpoints
 
-from .common import APACHE_VERSION, BASE_URL, CHECK_NAME, HERE, STATUS_CONFIG, STATUS_URL
+from .common import BASE_URL, CHECK_NAME, HERE, STATUS_CONFIG, STATUS_URL
 
 
 @pytest.fixture(scope="session")
@@ -38,15 +37,3 @@ def generate_metrics():
 @pytest.fixture
 def check():
     return lambda instance: Apache(CHECK_NAME, {}, [instance])
-
-
-@pytest.fixture(scope="session")
-def version_metadata():
-    major, minor, patch = APACHE_VERSION.split('.')
-    return {
-        'version.scheme': 'semver',
-        'version.major': major,
-        'version.minor': minor,
-        'version.patch': patch,
-        'version.raw': mock.ANY,
-    }