+++ cisco_aci/datadog_checks/cisco_aci/exceptions.py
@@ -17,7 +17,3 @@ class APIConnectionException(APIException):
 
 class APIParsingException(APIException):
     pass
-
-
-class ConfigurationException(Exception):
-    pass