+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/config.py
@@ -2,6 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import os
+from fnmatch import fnmatch
 
 import click
 import toml
@@ -24,6 +25,12 @@ from .console import CONTEXT_SETTINGS, echo_info, echo_success
 @click.group(context_settings=CONTEXT_SETTINGS, short_help='Manage the config file')
 def config():
     pass
+
+
+@config.command(context_settings=CONTEXT_SETTINGS, short_help='Open the config location in your file manager')
+def explore():
+    """Open the config location in your file manager."""
+    click.launch(CONFIG_FILE, locate=True)
 
 
 @config.command(context_settings=CONTEXT_SETTINGS, short_help='Show the location of the config file')
@@ -78,7 +85,7 @@ def set_value(ctx, key, value):
     """
     scrubbing = False
     if value is None:
-        scrubbing = key in SECRET_KEYS
+        scrubbing = any(fnmatch(key, pattern) for pattern in SECRET_KEYS)
         value = click.prompt('Value for `{}`'.format(key), hide_input=scrubbing)
 
     if key in ('core', 'extras', 'agent') and not value.startswith('~'):