+++ ibm_db2/tests/common.py
@@ -10,11 +10,14 @@ COMPOSE_FILE = os.path.join(HERE, 'docker', 'docker-compose.yaml')
 
 HOST = get_docker_hostname()
 PORT = '50000'
+DB = 'datadog'
+USERNAME = 'db2inst1'
+PASSWORD = 'db2inst1-pwd'
 
 CONFIG = {
-    'db': 'datadog',
-    'username': 'db2inst1',
-    'password': 'db2inst1-pwd',
+    'db': DB,
+    'username': USERNAME,
+    'password': PASSWORD,
     'host': HOST,
     'port': PORT,
     'tags': ['foo:bar'],