+++ snmp/tests/test_check.py
@@ -496,7 +496,7 @@ def test_cast_metrics(aggregator):
 def test_profile(aggregator):
     instance = common.generate_instance_config([])
     instance['profile'] = 'profile1'
-    init_config = {'profiles': {'profile1': {'definition': common.SUPPORTED_METRIC_TYPES}}}
+    init_config = {'profiles': {'profile1': {'definition': {'metrics': common.SUPPORTED_METRIC_TYPES}}}}
     check = SnmpCheck('snmp', init_config, [instance])
     check.check(instance)
 
@@ -512,7 +512,7 @@ def test_profile_by_file(aggregator):
     with temp_dir() as tmp:
         profile_file = os.path.join(tmp, 'profile1.yaml')
         with open(profile_file, 'w') as f:
-            f.write(yaml.safe_dump(common.SUPPORTED_METRIC_TYPES))
+            f.write(yaml.safe_dump({'metrics': common.SUPPORTED_METRIC_TYPES}))
         init_config = {'profiles': {'profile1': {'definition_file': profile_file}}}
         check = SnmpCheck('snmp', init_config, [instance])
         check.check(instance)
@@ -527,7 +527,28 @@ def test_profile_sys_object(aggregator):
     instance = common.generate_instance_config([])
     init_config = {
         'profiles': {
-            'profile1': {'definition': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.3.6.1.4.1.8072.3.2.10'}
+            'profile1': {
+                'definition': {'metrics': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.3.6.1.4.1.8072.3.2.10'}
+            }
+        }
+    }
+    check = SnmpCheck('snmp', init_config, [instance])
+    check.check(instance)
+
+    for metric in common.SUPPORTED_METRIC_TYPES:
+        metric_name = "snmp." + metric['name']
+        aggregator.assert_metric(metric_name, tags=common.CHECK_TAGS, count=1)
+    aggregator.assert_all_metrics_covered()
+
+
+def test_profile_sys_object_prefix(aggregator):
+    instance = common.generate_instance_config([])
+    init_config = {
+        'profiles': {
+            'profile1': {
+                'definition': {'metrics': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.3.6.1.4.1.8072.3.2.10'}
+            },
+            'profile2': {'definition': {'metrics': common.CAST_METRICS, 'sysobjectid': '1.3.6.1.4.*'}},
         }
     }
     check = SnmpCheck('snmp', init_config, [instance])
@@ -542,7 +563,9 @@ def test_profile_sys_object(aggregator):
 def test_profile_sys_object_unknown(aggregator):
     """If the fetched sysObjectID is not referenced by any profiles, check fails."""
     instance = common.generate_instance_config([])
-    init_config = {'profiles': {'profile1': {'definition': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.2.3.4.5'}}}
+    init_config = {
+        'profiles': {'profile1': {'definition': {'metrics': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.2.3.4.5'}}}
+    }
     check = SnmpCheck('snmp', init_config, [instance])
     check.check(instance)
 
@@ -561,28 +584,35 @@ def test_profile_sys_object_no_metrics():
 def test_discovery(aggregator):
     host = socket.gethostbyname(common.HOST)
     network = ipaddress.ip_network(u'{}/29'.format(host), strict=False).with_prefixlen
-    # Make sure the check handles bytes
-    network = network.encode("utf-8")
     check_tags = ['snmp_device:{}'.format(host)]
-    instance = {'name': 'snmp_conf', 'network_address': network, 'port': common.PORT, 'community_string': 'public'}
+    instance = {
+        'name': 'snmp_conf',
+        # Make sure the check handles bytes
+        'network_address': network.encode('utf-8'),
+        'port': common.PORT,
+        'community_string': 'public',
+    }
     init_config = {
         'profiles': {
-            'profile1': {'definition': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.3.6.1.4.1.8072.3.2.10'}
+            'profile1': {'definition': {'metrics': common.SUPPORTED_METRIC_TYPES, 'sysobjectid': '1.3.6.1.4.1.8072.*'}}
         }
     }
     check = SnmpCheck('snmp', init_config, [instance])
     try:
         for _ in range(30):
             check.check(instance)
-            if aggregator.metric_names:
+            if len(aggregator.metric_names) > 1:
                 break
             time.sleep(1)
+            aggregator.reset()
     finally:
         check._running = False
 
     for metric in common.SUPPORTED_METRIC_TYPES:
         metric_name = "snmp." + metric['name']
         aggregator.assert_metric(metric_name, tags=check_tags, count=1)
+
+    aggregator.assert_metric('snmp.discovered_devices_count', tags=['network:{}'.format(network)])
     aggregator.assert_all_metrics_covered()
 
 
@@ -652,7 +682,7 @@ def test_f5(aggregator):
     instance['community_string'] = 'f5'
     instance['enforce_mib_constraints'] = False
 
-    init_config = {'profiles': {'f5-big-ip': {'definition_file': path, 'sysobjectid': '1.3.6.1.4.1.3375.2.1.3.4.43'}}}
+    init_config = {'profiles': {'f5-big-ip': {'definition_file': path}}}
     check = SnmpCheck('snmp', init_config, [instance])
 
     check.check(instance)
@@ -670,23 +700,25 @@ def test_f5(aggregator):
         'sysTcpStatCloseWait',
         'sysTcpStatFinWait',
         'sysTcpStatTimeWait',
+        'sysUdpStatOpen',
+        'sysClientsslStatCurConns',
+    ]
+    counts = [
         'sysTcpStatAccepts',
         'sysTcpStatAcceptfails',
         'sysTcpStatConnects',
         'sysTcpStatConnfails',
-        'sysUdpStatOpen',
         'sysUdpStatAccepts',
         'sysUdpStatAcceptfails',
         'sysUdpStatConnects',
         'sysUdpStatConnfails',
-        'sysClientsslStatCurConns',
         'sysClientsslStatEncryptedBytesIn',
         'sysClientsslStatEncryptedBytesOut',
         'sysClientsslStatDecryptedBytesIn',
         'sysClientsslStatDecryptedBytesOut',
         'sysClientsslStatHandshakeFailures',
     ]
-    cpu_gauges = [
+    cpu_rates = [
         'sysMultiHostCpuUser',
         'sysMultiHostCpuNice',
         'sysMultiHostCpuSystem',
@@ -695,17 +727,39 @@ def test_f5(aggregator):
         'sysMultiHostCpuSoftirq',
         'sysMultiHostCpuIowait',
     ]
-    if_gauges = ['ifInOctets', 'ifInErrors', 'ifOutOctets', 'ifOutErrors']
+    if_gauges = ['ifAdminStatus', 'ifOperStatus']
+    if_counts = ['ifHCInOctets', 'ifInErrors', 'ifHCOutOctets', 'ifOutErrors']
     interfaces = ['1.0', 'mgmt', '/Common/internal', '/Common/http-tunnel', '/Common/socks-tunnel']
     for metric in gauges:
-        aggregator.assert_metric('snmp.{}'.format(metric), tags=common.CHECK_TAGS, count=1)
-    for metric in cpu_gauges:
-        aggregator.assert_metric('snmp.{}'.format(metric), tags=['cpu:0'] + common.CHECK_TAGS, count=1)
-        aggregator.assert_metric('snmp.{}'.format(metric), tags=['cpu:1'] + common.CHECK_TAGS, count=1)
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=common.CHECK_TAGS, count=1
+        )
+    for metric in counts:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=common.CHECK_TAGS, count=1
+        )
+    for metric in cpu_rates:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=['cpu:0'] + common.CHECK_TAGS, count=1
+        )
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=['cpu:1'] + common.CHECK_TAGS, count=1
+        )
+    for metric in if_counts:
+        for interface in interfaces:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric),
+                metric_type=aggregator.MONOTONIC_COUNT,
+                tags=['interface:{}'.format(interface)] + common.CHECK_TAGS,
+                count=1,
+            )
     for metric in if_gauges:
         for interface in interfaces:
             aggregator.assert_metric(
-                'snmp.{}'.format(metric), tags=['interface:{}'.format(interface)] + common.CHECK_TAGS, count=1
+                'snmp.{}'.format(metric),
+                metric_type=aggregator.GAUGE,
+                tags=['interface:{}'.format(interface)] + common.CHECK_TAGS,
+                count=1,
             )
     aggregator.assert_all_metrics_covered()
 
@@ -723,7 +777,7 @@ def test_router(aggregator):
 
     check.check(instance)
 
-    tcp_rates = [
+    tcp_counts = [
         'tcpActiveOpens',
         'tcpPassiveOpens',
         'tcpAttemptFails',
@@ -735,8 +789,8 @@ def test_router(aggregator):
         'tcpOutRsts',
     ]
     tcp_gauges = ['tcpCurrEstab']
-    udp_rates = ['udpHCInDatagrams', 'udpNoPorts', 'udpInErrors', 'udpHCOutDatagrams']
-    if_rates = [
+    udp_counts = ['udpHCInDatagrams', 'udpNoPorts', 'udpInErrors', 'udpHCOutDatagrams']
+    if_counts = [
         'ifInErrors',
         'ifInDiscards',
         'ifOutErrors',
@@ -750,7 +804,8 @@ def test_router(aggregator):
         'ifHCOutMulticastPkts',
         'ifHCOutBroadcastPkts',
     ]
-    ip_rates = [
+    if_gauges = ['ifAdminStatus', 'ifOperStatus']
+    ip_counts = [
         'ipSystemStatsHCInReceives',
         'ipSystemStatsHCInOctets',
         'ipSystemStatsInHdrErrors',
@@ -781,7 +836,7 @@ def test_router(aggregator):
         'ipSystemStatsHCInBcastPkts',
         'ipSystemStatsHCOutBcastPkts',
     ]
-    ip_if_rates = [
+    ip_if_counts = [
         'ipIfStatsHCInOctets',
         'ipIfStatsInHdrErrors',
         'ipIfStatsInNoRoutes',
@@ -812,21 +867,190 @@ def test_router(aggregator):
     ]
     for interface in ['eth0', 'eth1']:
         tags = ['interface:{}'.format(interface)] + common.CHECK_TAGS
-        for metric in if_rates:
-            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=tags, count=1)
-    for metric in tcp_rates:
-        aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=common.CHECK_TAGS, count=1)
+        for metric in if_counts:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+            )
+        for metric in if_gauges:
+            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=tags, count=1)
+    for metric in tcp_counts:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=common.CHECK_TAGS, count=1
+        )
     for metric in tcp_gauges:
         aggregator.assert_metric(
             'snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=common.CHECK_TAGS, count=1
         )
-    for metric in udp_rates:
-        aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=common.CHECK_TAGS, count=1)
+    for metric in udp_counts:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=common.CHECK_TAGS, count=1
+        )
     for version in ['ipv4', 'ipv6']:
         tags = ['ipversion:{}'.format(version)] + common.CHECK_TAGS
-        for metric in ip_rates:
-            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=tags, count=1)
-        for metric in ip_if_rates:
+        for metric in ip_counts:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+            )
+        for metric in ip_if_counts:
             for interface in ['17', '21']:
                 tags = ['ipversion:{}'.format(version), 'interface:{}'.format(interface)] + common.CHECK_TAGS
-                aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.RATE, tags=tags, count=1)
+                aggregator.assert_metric(
+                    'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+                )
+
+    aggregator.assert_all_metrics_covered()
+
+
+def test_f5_router(aggregator):
+    instance = common.generate_instance_config([])
+    # We need the full path as we're not in installed mode
+    path = os.path.join(os.path.dirname(snmp.__file__), 'data', 'profiles', 'generic-router.yaml')
+
+    # Use the generic profile against the f5 device
+    instance['community_string'] = 'f5'
+    instance['profile'] = 'router'
+    instance['enforce_mib_constraints'] = False
+
+    init_config = {'profiles': {'router': {'definition_file': path}}}
+    check = SnmpCheck('snmp', init_config, [instance])
+
+    check.check(instance)
+
+    if_counts = [
+        'ifInErrors',
+        'ifInDiscards',
+        'ifOutErrors',
+        'ifOutDiscards',
+        'ifHCInOctets',
+        'ifHCInUcastPkts',
+        'ifHCInMulticastPkts',
+        'ifHCInBroadcastPkts',
+        'ifHCOutOctets',
+        'ifHCOutUcastPkts',
+        'ifHCOutMulticastPkts',
+        'ifHCOutBroadcastPkts',
+    ]
+    if_gauges = ['ifAdminStatus', 'ifOperStatus']
+    # We only get a subset of metrics
+    ip_counts = [
+        'ipSystemStatsHCInReceives',
+        'ipSystemStatsInHdrErrors',
+        'ipSystemStatsOutFragReqds',
+        'ipSystemStatsOutFragFails',
+        'ipSystemStatsHCOutTransmits',
+        'ipSystemStatsReasmReqds',
+        'ipSystemStatsHCInMcastPkts',
+        'ipSystemStatsReasmFails',
+        'ipSystemStatsHCOutMcastPkts',
+    ]
+    interfaces = ['1.0', 'mgmt', '/Common/internal', '/Common/http-tunnel', '/Common/socks-tunnel']
+    for interface in interfaces:
+        tags = ['interface:{}'.format(interface)] + common.CHECK_TAGS
+        for metric in if_counts:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+            )
+        for metric in if_gauges:
+            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=tags, count=1)
+    for version in ['ipv4', 'ipv6']:
+        tags = ['ipversion:{}'.format(version)] + common.CHECK_TAGS
+        for metric in ip_counts:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+            )
+
+    aggregator.assert_all_metrics_covered()
+
+
+def test_3850(aggregator):
+    instance = common.generate_instance_config([])
+    # We need the full path as we're not in installed mode
+    path = os.path.join(os.path.dirname(snmp.__file__), 'data', 'profiles', 'cisco-3850.yaml')
+    instance['community_string'] = '3850'
+    instance['profile'] = 'cisco-3850'
+    instance['enforce_mib_constraints'] = False
+
+    init_config = {'profiles': {'cisco-3850': {'definition_file': path}}}
+    check = SnmpCheck('snmp', init_config, [instance])
+
+    check.check(instance)
+
+    tcp_counts = [
+        'tcpActiveOpens',
+        'tcpPassiveOpens',
+        'tcpAttemptFails',
+        'tcpEstabResets',
+        'tcpHCInSegs',
+        'tcpHCOutSegs',
+        'tcpRetransSegs',
+        'tcpInErrs',
+        'tcpOutRsts',
+    ]
+    tcp_gauges = ['tcpCurrEstab']
+    udp_counts = ['udpHCInDatagrams', 'udpNoPorts', 'udpInErrors', 'udpHCOutDatagrams']
+    if_counts = ['ifInErrors', 'ifInDiscards', 'ifOutErrors', 'ifOutDiscards']
+    ifx_counts = [
+        'ifHCInOctets',
+        'ifHCInUcastPkts',
+        'ifHCInMulticastPkts',
+        'ifHCInBroadcastPkts',
+        'ifHCOutOctets',
+        'ifHCOutUcastPkts',
+        'ifHCOutMulticastPkts',
+        'ifHCOutBroadcastPkts',
+    ]
+    if_gauges = ['ifAdminStatus', 'ifOperStatus']
+    # We're not covering all interfaces
+    interfaces = ["GigabitEthernet1/0/{}".format(i) for i in range(1, 48)]
+    for interface in interfaces:
+        tags = ['interface:{}'.format(interface)] + common.CHECK_TAGS
+        for metric in if_counts:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+            )
+        for metric in if_gauges:
+            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=tags, count=1)
+    interfaces = ["Gi1/0/{}".format(i) for i in range(1, 48)]
+    for interface in interfaces:
+        tags = ['interface:{}'.format(interface)] + common.CHECK_TAGS
+        for metric in ifx_counts:
+            aggregator.assert_metric(
+                'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1
+            )
+    for metric in tcp_counts:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=common.CHECK_TAGS, count=1
+        )
+    for metric in tcp_gauges:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=common.CHECK_TAGS, count=1
+        )
+    for metric in udp_counts:
+        aggregator.assert_metric(
+            'snmp.{}'.format(metric), metric_type=aggregator.MONOTONIC_COUNT, tags=common.CHECK_TAGS, count=1
+        )
+    sensors = [1006, 1007, 1008, 2006, 2007, 2008]
+    for sensor in sensors:
+        tags = ['sensor_id:{}'.format(sensor), 'sensor_type:8'] + common.CHECK_TAGS
+        aggregator.assert_metric('snmp.entSensorValue', metric_type=aggregator.GAUGE, tags=tags, count=1)
+    fru_metrics = ["cefcFRUPowerAdminStatus", "cefcFRUPowerOperStatus", "cefcFRUCurrent"]
+    frus = [1001, 1010, 2001, 2010]
+    for fru in frus:
+        tags = ['fru:{}'.format(fru)] + common.CHECK_TAGS
+        for metric in fru_metrics:
+            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=tags, count=1)
+
+    cpus = [1000, 2000]
+    cpu_metrics = ["cpmCPUTotalMonIntervalValue", "cpmCPUMemoryUsed", "cpmCPUMemoryFree"]
+    for cpu in cpus:
+        tags = ['cpu:{}'.format(cpu)] + common.CHECK_TAGS
+        for metric in cpu_metrics:
+            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=tags, count=1)
+    cie_metrics = ["cieIfLastInTime", "cieIfLastOutTime", "cieIfInputQueueDrops", "cieIfOutputQueueDrops"]
+    for interface in interfaces:
+        tags = ['interface:{}'.format(interface)] + common.CHECK_TAGS
+        for metric in cie_metrics:
+            aggregator.assert_metric('snmp.{}'.format(metric), metric_type=aggregator.GAUGE, tags=tags, count=1)
+        aggregator.assert_metric('snmp.cieIfResetCount', metric_type=aggregator.MONOTONIC_COUNT, tags=tags, count=1)
+
+    aggregator.assert_all_metrics_covered()