+++ supervisord/tests/common.py
@@ -15,6 +15,8 @@ PORT = 19001
 HERE = os.path.dirname(os.path.abspath(__file__))
 URL = "http://{}:{}".format(HOST, PORT)
 
+SUPERVISOR_VERSION = os.getenv('SUPERVISOR_VERSION')
+
 # Supervisord should run 3 programs for 10, 20 and 30 seconds
 # respectively.
 # The following dictionnary shows the processes by state for each iteration.