+++ gunicorn/CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG - gunicorn
 
+## 1.8.1 / 2019-12-13
+
+* [Fixed] Bump psutil to 5.6.7. See [#5210](https://github.com/DataDog/integrations-core/pull/5210).
+
+## 1.8.0 / 2019-12-02
+
+* [Fixed] Remove shlex. See [#5064](https://github.com/DataDog/integrations-core/pull/5064).
+* [Fixed] Upgrade psutil dependency to 5.6.5. See [#5059](https://github.com/DataDog/integrations-core/pull/5059).
+* [Added] Add version metadata. See [#4968](https://github.com/DataDog/integrations-core/pull/4968).
+
 ## 1.7.2 / 2019-10-11
 
 * [Fixed] Upgrade psutil dependency to 5.6.3. See [#4442](https://github.com/DataDog/integrations-core/pull/4442).