+++ process/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG - process
 
+## 1.10.3 / 2019-12-13
+
+* [Fixed] Bump psutil to 5.6.7. See [#5210](https://github.com/DataDog/integrations-core/pull/5210).
+
+## 1.10.2 / 2019-12-02
+
+* [Fixed] Upgrade psutil dependency to 5.6.5. See [#5059](https://github.com/DataDog/integrations-core/pull/5059).
+
 ## 1.10.1 / 2019-10-11
 
 * [Fixed] Make iowrite_bytes and ioread_bytes counts. See [#4655](https://github.com/DataDog/integrations-core/pull/4655).